voxcpm-tts-inference/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
pydantic>=2.5
soundfile>=0.12
numpy>=1.24
pyyaml>=6.0
python-dotenv>=1.0
